feat: 게시글 삭제 기능

+++ src/main/java/hello/moviecomm/member/domain/Authority.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class Authority {
-    private final Long authorityCode;
+    private final Integer authorityCode;
     private final String name;
     private final String description;
 }
